backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai>=0.8.0
python-dotenv==1.0.1
pydantic==2.8.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
